requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0
gspread==6.1.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.41.0
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.82.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9
selenium==4.28.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.41.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
websocket-client==1.8.0
wsproto==1.2.0
langchain_community
